requirements.txt
Flask>=2.0
requests>=2.20
google-generativeai>=0.5
python-dotenv>=0.15
